requirements.txt
asgiref==3.5.1
attrs==21.4.0
Django==4.0.4
django-filter==21.1
django-safedelete==1.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
importlib-metadata==4.11.3
iniconfig==1.1.1
Markdown==3.3.7
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
routers==0.10.1
sqlparse==0.4.2
tomli==2.0.1
zipp==3.8.0
